pays/zarinpal/requirements.txt
python-telegram-bot==13.14
peewee
psycopg2
coloredlogs
jdatetime
suds-py3
requests
qrcode
Pillow